Added unpaddedSize_ to Tensor (#2970)
Summary:
This PR adds a new field to Tensor: unpaddedSize_.
Since we have a static batch size we copy the entire tensor to the device regardless of number of inputs. This would allow a DeviceManager to only copy the inputs if it supports it.
This means adding additional metaData to Tensor, this seemed cleaner than passing side channel data.

Documentation: NA
Pull Request resolved: https://github.com/pytorch/glow/pull/2970

Differential Revision: D15577757

Pulled By: gcatron

fbshipit-source-id: 105bad588c13cd6746a63e14594c0ba57c8fb974