Improve copy elimination optimisation (#2935)
Summary:
**Description**
This commit improves the copy instruction elimination optimization in
the case that the src and dest operands are `tensorview` instructions.
Prior to this commit, this optimization used to eliminate copy
instructions if the src an dest `tensorview` offsets were equal. This does
not work for all cases; for example, offsets {0} and {0, 0} are not
programmatically equal, but are semantically equal.

This commit modifies this optimization to check that the linearized
offsets into the underlying buffers after peeling away all tensorviews
are equal before erasing the copy instruction. For example, the previous
implementation incorrectly eliminated a copy between identical offsets
into two different intermediate tensorviews that provide a view into
different sections of the same buffer. This commit addresses this.

This commit also modifies `verify()` for `Insert/ExtractTensorInst` and
`TensorViewInst` to check that the number of offsets in the instruction is equal to the size of the
src/dest shape. This simple check precludes the case in which offsets `{0}` and `{0,
0}` are compared by this optimization and determined not to be equal.
Code that creates these types of instructions has also been updated
accordingly.

**Testing**
This commit adds the case mentioned above with a copy between identical
offsets into two separate `tensorviews` that have different offsets into
the same underlying buffer as a unit test in IROptTest. All unit tests pass.
Pull Request resolved: https://github.com/pytorch/glow/pull/2935

Differential Revision: D15424891

Pulled By: SplitInfinity

fbshipit-source-id: e37bbca70b2a980f6d5c8550c7f3cb570f9b45bc