[Interpreter] Implement peer-to-peer communication primitives
Description:
This commit implements peer-to-peer communication for the Interpreter
backend. This feature can be used to send/receive Tensors from a
function on one device to/from a second function on another device
without using host memory as an intermediary (which is the status quo).

This commit adds Send and Receive nodes to the graph IR to explictly
represent these operations as well as corresponding send and receive
instructions in the instruction IR. A SendNode and ReceiveNode are
"paired" at execution time using the Tensor bound to their "address"
inputs. This is a Tensor of shape {1} that contains some sort of token
or address that the backend can interpret however it wants to transmit
data between the Send and Receive nodes to which that token is bound.
In addition to sending and receiving data using peer-to-peer
communication, the Send and Receive nodes also write the data they
send/receive to a Placeholder so that they can be inspected (e.g. for
debugging).

In the case of the Interpreter backend, the "token" mentioned above is
the address of the ReceiveNode's Placeholder, and the data is
transferred using a future-promise pair of type Tensor (i.e.
std::future<Tensor> and std::promise<Tensor>). Each tuple of (function
name, placeholder address) maps to a future-promise pair, and the
Interpreter implementation of send fulfills the promise, and the
implementation of recv waits on the future.

Testing:
This commit adds a new unit test, PeerToPeerTest, which tests
sending/receiving data with one sender + one receiver, one sender with
multiple send nodes and multiple receivers, and multiple senders and
multiple receivers. All tests pass.

Documentation:
Once the design is finalized after a few rounds of code review,
documentation will be added to reflect the updated design.