[PlaceholderBindings] Add DCHECK to insert to check that types match
Description:
This commit adds a debug-mode check to `PlaceholderBindings::insert`
that checks that the types of the `Placeholder` and `Tensor` passed to
the method match. It is currently possible to bind a `Placeholder`
to any `Tensor` and that seems like a recipe for disaster. It will still
be possible after this commit, but will crash if the project is built in
Debug mode.

Testing:
All unit tests pass.