Add DCHECK to insert to check that types match (#3042)
Summary:
**Description**
This commit adds a debug-mode check to `PlaceholderBindings::insert`
that checks that the types of the `Placeholder` and `Tensor` passed to
the method match. It is currently possible to bind a `Placeholder`
to any `Tensor` and that seems like a recipe for disaster. It will still
be possible after this commit, but will crash if the project is built in
Debug mode.

**Test Plan**
All unit tests pass.
Pull Request resolved: https://github.com/pytorch/glow/pull/3042

Differential Revision: D15632584

Pulled By: SplitInfinity

fbshipit-source-id: a63437b490066a564d6689101f75511e6978a848