Introduce generation of Constant Folding methods in Node Builder
This patch creates infrastructure that allows to constant-fold nodes early of
load, if their inputs are all constant, rather than leave it up to backend to
deal with them later. By doing so, we reduce the size of graph and enable more
models, because some nodes require that some of their inputs are constants.

Theoretically, Constant Folding can be done for all nodes. But this API allows
to implement it for particular nodes only. Two reasons for that:
- We can implement constant-folding incrementally and not make huge changes in
  code;
- We might not want to evalutate some of the nodes on load because of
  performance reasons.

We can then add the constant folding to the nodes one by one as we need it.