Introduce Early Constant Folding for Gather node
This patch implements early constant folding for Gather node with all
constant inputs.

The motivating example of this change is following:

  Reshape(Gather(Shape(X))).

In Glow, Shape is always a constant, and Reshape requires its Shape argument to
be constant. If we constant-fold Gather on load and not postpone it for backend,
we can load this model.