Add verbose mode to isEqual check for easier debugging. (#3054)
Summary:
Add a verbose mode to isEqual so that when there is a mismatch during our unit tests we log info the error. I disabled this in the two places we aren't using this in unit tests and so we want to early exit when Tensors aren't equal, i.e. GraphOptimizer and PlaceholderBindings.
Pull Request resolved: https://github.com/pytorch/glow/pull/3054

Differential Revision: D15702217

Pulled By: jfix71

fbshipit-source-id: 2d2d419b84b7445e60955bbce2a9e96a8edbdaa0