Remove the Executor factory (#3070)
Summary:
Currently we have a factory method that creates Executors in the runtime - similar to Backends and DeviceManagers. However we only have one type of Executor, and I don't see us on a trend to needing multiple types you can drop in different scenarios. The factory approach also makes it harder to have different configurations for different executors (we'd need a DeviceConfig equivalent for Executors, probably).

So proposing to remove it and the associated indirection until a future point where we feel like we need it.

Documentation: n/a
Pull Request resolved: https://github.com/pytorch/glow/pull/3070

Reviewed By: arunm-git

Differential Revision: D15747051

Pulled By: nickgg

fbshipit-source-id: 26f3b837686e03a18886023cec4766612c5cd43b