Add setup configuration to the HostManager (#3078)
Summary:
The Runtime has various configuration parameters that there is no mechanism to set (such as the number of threads in the ThreadPoolExecutor). This PR adds a config structure to HostManager allowing passing these parameters in and implements both the number of threads and the limit of active requests.

Documentation: just comments
Pull Request resolved: https://github.com/pytorch/glow/pull/3078

Differential Revision: D15768990

Pulled By: nickgg

fbshipit-source-id: a26acdac0623247e3814457ab6a00657b8c6e815