[TraceEvents][OpenCL] Fix OpenCL trace output
Description:
This commit ensures that traces generated by the OpenCL backend will
render correctly regardless of the default time duration unit returned
by std::chrono::system_clock().now(). This function is used in
TraceEvents::now() to get the current timestamp but it is cast to
std::chrono::microseconds before it is returned. This commit modifies
the call to this function in OpenCLFunction::translateTraceEvents to
also cast the result to std::chrono::microseconds. Without this, the
events corresponding to OpenCL kernels do not render currently as being
nested inside events that should encompass them (e.g.
DeviceManager::runFunction) if the return value of std::chrono::system_clock().now() is not in microseconds.

Testing:
Visually examined traces on two machines, one which returns the current
time in microseconds and another that returns it in nanoseconds.