[OpenCL] Fix bugs in BatchedReduceAddInst implementation
Description:
This commit fixes two bugs in the OpenCL implementation of
BatchedReduceAddInst and adds a few comments for clarity.

The first is a segmentation fault caused by
incorporating feedback on #2958. A suggestion was made to make the loop
variable i in the loop that computes batchSliceSizes count down instead of
count up, but this was done without changing the type (which was size_t,
an unsigned type), so the loop never terminates and eventually leads to a
segmentation fault.

The second bug is an incorrect computation of destSliceSizes. Instead of
multiplying the slice size at a dimension with the number of elements in
that same dimension, the code was multiplying the former with the number
of elements in the *adjacent* dimension. This was surfaced by the unit
test added in #2958 for axis = 2.

Testing:
1) ninja check with OpenCL enabled, DEBUG mode
2) ninja check with OpenCL enabled, RELEASE mode
3) ninja check with OpenCL enabled, ASAN+UBSAN mode