[runtime] Perform a round of backend-independent optimizations before partitioning
Doing so reduces the number of nodes to be partitioned and partitioner has less work to do.

When using a custom experimental backend and a heterogeneous partitioning, I've seen the number of partitions produced for Glow's ResNet50 model going down from 101 to 3 after running a round of backend-independent optimizations before partitioning.