Log the change to the input(operand) of a node. (#3111)
Summary:
Log the change to the input(operand) of a node. This is for the case when the input of one node changes after the creation, which may happen at:
- transformForPrecisionMode
  - quantizeFunction(): via invocations to setNthInput()
  - convertToFP16: via invocations to setOperand() and setOperand()
- several other trivial cases like changing the bias of a Conv node
  - via invocation to setNthInput()

This PR intercepts the implementation of setNthInput() and the certain caller(inside FunctionConverter) to setOperand(), such that the changes to the input of a node is caught and logged.
Pull Request resolved: https://github.com/pytorch/glow/pull/3111

Test Plan:
test with Quantization.TestProfileQuantizationOfUnloweredFC, and the log output is attached below:
![Screen Shot 2019-06-13 at 9 37 32 PM](https://user-images.githubusercontent.com/8838608/59483779-cb75b800-8e23-11e9-80a4-3c3019088c2c.png)

Differential Revision: D15830476

Pulled By: ZchiPitt

fbshipit-source-id: 4cb643c968cb2a58d423b350f2b5b69276980a11