[OpenCL] Add TraceEvent for compiling custom convolution kernels
Description:
This commit adds a TraceEvent that covers the time spent during
OpenCLFunction::execute() on compiling custom versions of
convolution kernels according to the parameters of the particular
convolutions present in the function. Depending on the OpenCL
implementation, this time can manifest as mysterious large gaps in the trace file.

Testing:
Generated traces for tracing-compare example.