Kill glow::onnxifi::Backend
Summary:
`Backend` is purely a wrapper around `BackendId` and it seems to exist
only because the onnxifi API likes to talk about "backendId" and "backend" like
they're different things. But I think it makes the code clearer if we just make
them one class and simply hide this fact from onnxifi.

Differential Revision: D15903689

fbshipit-source-id: c0c2205a057eaca616508ecba701be6bb7367a09