Kill glow::onnxifi::Backend
Summary:
`Backend` is purely a wrapper around `BackendId` and it seems to exist
only because the onnxifi API likes to talk about "backendId" and "backend" like
they're different things. But I think it makes the code clearer if we just make
them one class and simply hide this fact from onnxifi.

Reviewed By: jackm321, rdzhabarov

Differential Revision: D15903689

fbshipit-source-id: 840c0728a94993fa27189ede8eac4c9a92ae16c7