Avoid zeroing Tensors unnecessarily (#3128)
Summary:
We zero tensors whenever they are `reset()`, and we reset Tensors a lot. Instead we should only zero buffers when we really need to. An example is if we're loading a constant from the protobuf into a Tensor it doesn't matter what the original value was.

Many tests broke when I did this, and most of the time I fixed it by intentionally calling `zero()` on the Tensor. Usually it was just because the test expected the result of working on an uninitialized Tensor to be the same between runs (e.g. when doing a comparison against the Interpreter).

A few cases were legitimate bugs hidden by the fact the Tensors happened to be zero, which i've fixed here:
* Quantization Profiling: The Tensor holding the profile is modified in place and needs to be manually zeroed. In this diff I've done it by marking it as "Trainable" which may be slight abuse of the term (we're training the quantizer to choose good params... ?).
* The Interpreters CrossEntropyLoss calculation for floats was modifying its output Tensor without clearing it first.
* Training: When we lowered the SGD node we weren't setting its PH to be trainable. If we guarantee all trainable Tensors are zeroed this works as intended. (Question: should training Tensors always start zero? Isn't it normal to start in a random state?).
* The `inferNonSquarePaddingConv` test helper used the wrong dimensions.
* ThreadPoolExecutor: This was the hardest to work out. The test was incorrectly verifying that output Tensors were consistent when passed into the DeviceManager but that is definitely not true now that we use a TensorPool, and was only passing because Tensors were initialized zero and were never modified (so all Tensors were the same).

Not fixed but weird:
* inferConvDKKC8 test is using the wrong indexes I think, will call out in a comment.

Documentation: Fixes https://github.com/pytorch/glow/issues/3037
Pull Request resolved: https://github.com/pytorch/glow/pull/3128

Test Plan:
unit tests under Release and ASAN, Tested again with the GLOW_DEBUG_TENSOR_INIT define set to test when the initial state is definitely not 0.

Manual testing with resnet-runtime shows a **90% reduction in zeroing buffers**:

before:
```
│I0617 21:46:29.172291 451777984 resnet-runtime.cpp:221] resets: 505
│I0617 21:46:29.172297 451777984 resnet-runtime.cpp:222] zeroes: 558

```
after:
```
│I0617 21:43:12.518196 208471488 resnet-runtime.cpp:221] resets: 505
│I0617 21:43:12.518205 208471488 resnet-runtime.cpp:222] zeroes: 53

```

(I determined this with a global atomic, counting code is not in the diff)

Will follow up later with some more prod like tests, but there are decisions here worth getting comment on early.

Reviewed By: bertmaher

Differential Revision: D15886809

Pulled By: nickgg

fbshipit-source-id: 0f4a973edc1c11e35c82dedbb017a9533d4f998c