Add TraceEvent for compiling custom convolution kernels (#3140)
Summary:
**Description**
This commit adds a TraceEvent that covers the time spent during
OpenCLFunction::execute() on compiling custom versions of
convolution kernels according to the parameters of the particular
convolutions present in the function. Depending on the OpenCL
implementation, this time can manifest as mysterious large gaps in the trace file.

**Test Plan**
Generated traces for `tracing-compare` example with AMD OpenCL.

Without this PR:
![Screen Shot 2019-06-19 at 3 27 16 PM](https://user-images.githubusercontent.com/4392003/59805729-f1c5a880-92a6-11e9-9212-c4050302234d.png)

With this PR:
![Screen Shot 2019-06-19 at 3 21 26 PM](https://user-images.githubusercontent.com/4392003/59805740-f8542000-92a6-11e9-91e6-f17290622763.png)
Pull Request resolved: https://github.com/pytorch/glow/pull/3140

Differential Revision: D15908094

Pulled By: SplitInfinity

fbshipit-source-id: f787d24470f851b02210c6a74d328b109fd177ff