Support int32 and Boolean types for Concat in Interpreter (#3155)
Summary:
This patch adds support for Concat and related nodes for int32 and Boolean types
in interpreter. The main purpose of this is to leverage constant-folding to get rid of
such tensors, that's why no CPU support added so far (will be done if needed).
Pull Request resolved: https://github.com/pytorch/glow/pull/3155

Test Plan: Added unit tests

Differential Revision: D15946617

Pulled By: jfix71

fbshipit-source-id: 6acf5116d2e250f9c5b4a12b6b892e7d593c22d5