[OpenCL] Optimize BatchedReduceAdd implementation
Description:
This commit reduces the time taken to execute a batchedreduceadd
instruction in the OpenCL backend by moving the transfer of the input
and output slice size data from execution time to addNetwork time.

The slice sizes of the input and output can be computed using static
shape information at compile time, so they don't have to be computed and
transferred once per operation invocation at runtime. This commit introduces
a post-lowering OCLBackend transformation that replaces BatchedReduceAddNode
with a semantically identical OCLBatchedReduceAddNode that has two additional
inputs for the slice sizes of the input and output nodes. The code to
compute these slice sizes has been moved from OpenCLFunction::execute()
to OCLBackend::transformPostLowering and modified to write into the
payload Tensors of Constants that are used as the inputs to the
previously mentioned OCLBatchedReduceAddNode. These slice sizes are then
copied to the device with the rest of the constants needed by the
function.

Test Plan:
All unit tests pass.