Support casts i32<->i64<->float<->float16 in Interpreter (#3179)
Summary:
The main purpose of that is constant-folding, that is why only the interpreter
backend is supported so far. Can be done for CPU backend separately if needed.
Pull Request resolved: https://github.com/pytorch/glow/pull/3179

Test Plan:
Added unit tests.

Please see a detailed explanation of how to fill out the fields in the relevant sections in PULL_REQUEST.md.

Differential Revision: D16016739

Pulled By: rdzhabarov

fbshipit-source-id: 92eb37df000d31cc03b1010b431db9073d74a14b