Truncate 64-bit indices to 32 (#3189)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/3189

We've found that PCIe transfers are a bottleneck for networks that do
a large number of embedding lookups.  Part of the problem is that these indices
are often unnecessarily specified as int64 due to ML framework defaults.
Ideally we'd fix the framework, but as a workaround we forcibly truncate these indices to 32 bits.

Reviewed By: jackm321, rdzhabarov

Differential Revision: D15609734

fbshipit-source-id: 4853d983c3971c63a905ad612901be49c8f4130e