Optimize BatchedReduceAdd implementation (#3190)
Summary:
**Description**
This commit reduces the time taken to execute a `batchedreduceadd`
instruction in the OpenCL backend by moving the transfer of the input
and output slice size data from execution time to `addNetwork` time.

The slice sizes of the input and output can be computed using static
shape information at compile time, so they don't have to be computed and
transferred once per operator invocation at runtime. This commit introduces
a post-lowering `OCLBackend` transformation that replaces `BatchedReduceAddNode`
with a semantically identical `OCLBatchedReduceAddNode` that has two additional
inputs for the slice sizes of the input and output nodes. The code to
compute these slice sizes has been moved from `OpenCLFunction::execute`
to `OCLBackend::transformPostLowering` and modified to write into the
payload `Tensors` of `Constants` that are used as the inputs to the
previously mentioned `OCLBatchedReduceAddNode`. These slice sizes are then
copied to the device with the rest of the constants needed by the
function.

**Test Plan**
All unit tests pass.

DLRM trace before this optimization:
![Screen Shot 2019-06-27 at 2 57 36 PM](https://user-images.githubusercontent.com/4392003/60303810-43e07c80-98ec-11e9-9b11-d1f73694d2be.png)

DLRM trace after this optimization:
![Screen Shot 2019-06-27 at 2 57 18 PM](https://user-images.githubusercontent.com/4392003/60303814-493dc700-98ec-11e9-9481-d0c12a8e1362.png)

Time taken to process one minibatch has decreased by 20%.
Pull Request resolved: https://github.com/pytorch/glow/pull/3190

Differential Revision: D16038375

Pulled By: SplitInfinity

fbshipit-source-id: 4767f4db42de793635969236fd9f2b328946842a