Fix optimizeConversions: consider cases where cast changes value
The optimizing pass optimizeConversions assumes that the following operation
is always NOOP:

  T1 X = ...;
  T1 Y = T1(T2(X));

and replaces Y with X. In fact, there are at least two cases when such
replacement is invalid:
  - if T1 is a floating-point type and T2 is not, T2(X) may change the value by
    dropping the fractional part;
  - if mantissa of T2 has less significant bits than mantissa of T1, this cast
    may lead to loss of senior significant bits.

In both cases, the said replacement is non-equivalent. Justification that we can
can still perform this transform despite the sematic change because we eliminate
precision loss does not hold, because transform like float -> int -> float
can be done intentionally to strip away fractional part, and this transform
removes a semantically important piece.

This patch fixes that situation.