[RecommendationSystemTest] Enable tracing
Description:
This commit enables tracing for RecommendationSystemTest. The traces are
saved to files in the directory specified by the -trace-dir option. The
unique name for each test is obtained by concatenating the
name of its containing test case as well as its name. This commit also
adds code to print an error message if the trace file cannot be
opened for whatever reason.

Testing:
Ran the test with and without the option. With the option, a trace is
dumped for each test and can be viewed with chrome://tracing. Without
the option, no traces are dumped.