Added new BackendHints object (#3234)
Summary:
This PR adds a new BackendHints object and adds it to BackendOptions and DAGNode.
This PR also adds the plumbing for BackendHints created by the Partitioner to be piped through the Provisioner to the Backend.
Currently there are two hints:
SRAM Pinning - An ordered list of Storage nodes to be pinned to SRAM. Order denotes priority.
Resource Reservation - An unsigned integer for the number of execution units to use for the function.

The plan is that new passes in the Partitioner will determine these hints and populate the backendHints field of the DAGNode for the function. This is then passed to the backend by the Provisioner.

Documentation: Once the design is agreed on, Backends.md will need to be updated discussing the hints.

Related to https://github.com/pytorch/glow/issues/3235
Pull Request resolved: https://github.com/pytorch/glow/pull/3234

Test Plan: ninja test

Differential Revision: D16345864

Pulled By: gcatron

fbshipit-source-id: 1d78e30d8480989d37b382f44a9f03db416cbacd