Fix a bug: quantization profiling flow with partitioning can only be enabled when there is no previous registered network in HM. (#3268)
Summary:
When we enabled quantization profiling flow in Partitioner, we forced Provisioner to be reset. If there are registered networks in HM already, the status will be lost. Therefore, we restrict the quantization profiling flow to only 1 network.

Documentation:

[Optional Fixes #issue]
Pull Request resolved: https://github.com/pytorch/glow/pull/3268

Test Plan:
ninja test, ./tests/images/run.sh
Please see a detailed explanation of how to fill out the fields in the relevant sections in PULL_REQUEST.md.

Differential Revision: D16413045

Pulled By: beicy

fbshipit-source-id: 48ec69563894381ec69f5f7398a7c730c382082b