[RuntimeBundle] Disable implicit copy for RuntimeBundle
Summary:
This commit disables implicit copying of RuntimeBundles. They should be
moved whenever possible and explicitly copied if moving is not possible.
An example of this is in the Provisioner, where the RuntimeBundle is
copied from a CompiledFunction into is encompassing DAGNode. This commit
also adds an isValid_ flag to the RuntimeBundle and sets it to false
after an instance has been moved out of in order to help debug problems
arising from manipulating a stale RuntimeBundle instance.

Test Plan:
ninja check