Disable implicit copy for RuntimeBundle (#3278)
Summary:
**Summary**
This commit disables implicit copying of `RuntimeBundles`. They should be
moved whenever possible and explicitly copied if moving is not possible.
An example of this is in the `Provisioner`, where the `RuntimeBundle` is
copied from a `CompiledFunction` into its encompassing `DAGNode`. This commit
also adds an `isValid_` flag to the `RuntimeBundle` and sets it to false
after an instance has been moved out of in order to help debug problems
arising from manipulating a stale `RuntimeBundle` instance.

**Test Plan**
`ninja check`

**Fixes**
This PR fixes https://github.com/pytorch/glow/issues/3259.
Pull Request resolved: https://github.com/pytorch/glow/pull/3278

Differential Revision: D16442341

Pulled By: SplitInfinity

fbshipit-source-id: b54aaee3d3e99071c7956c3bcc0783282078897c