[PyTorch] Add support for exponential operator
Description:
This commit adds support for the exponential function operator to the
PyTorchModelLoader so that torch.exp can be executed on Glow.

Testing:
This commit adds a unit test that runs a graph with two exp nodes on Glow and
the PyTorch CPU backend and checks that the results are identical.