Add support for exponential operator (#3290)
Summary:
**Summary**
This commit adds support for the exponential function operator to the
PyTorchModelLoader so that torch.exp can be executed on Glow.

**Test Plan**
This commit adds a unit test that runs a graph with two exp nodes on Glow and
the PyTorch CPU backend and checks that the results are identical.
Pull Request resolved: https://github.com/pytorch/glow/pull/3290

Differential Revision: D16477743

Pulled By: SplitInfinity

fbshipit-source-id: d0d9bde52c4b6b8a0747ace3cee32774c5561b79