[PyTorch] Add support for elementwise max operator
Summary:
This commit adds support for the elementwise max operator to the
PyTorchModelLoader so that torch.max(a, b) can be executed on Glow.

Test Plan:
This commit adds a unit test that runs a graph with two max nodes on Glow and
the PyTorch CPU backend and checks that the results are identical.