[PyTorch] Add support for sqrt operator
Summary:
This commit adds support for the elementwise sqrt operator to the
PyTorchModelLoader so that torch.sqrt(a) can be executed on Glow.

Test Plan:
This commit adds a unit test that runs a graph with two sqrt nodes on Glow and
the PyTorch CPU backend and checks that the results are identical.