Add support for elementwise max operator (#3303)
Summary:
**Summary**
This commit adds support for the elementwise max operator to the
PyTorchModelLoader so that `torch.max(a, b)` can be executed on Glow.

**Test Plan**
This commit adds a unit test that runs a graph with two max nodes on Glow and
the PyTorch CPU backend and checks that the results are identical.
Pull Request resolved: https://github.com/pytorch/glow/pull/3303

Differential Revision: D16511532

Pulled By: SplitInfinity

fbshipit-source-id: a91809a1957020765222fe8befecde80643bb401