Add FP16 versions of RWQ-SLWS (#3265)
Summary:
Add support for FP16 RWQ-SLWS, fused and unfused. In this version of the op, the per-row scale/offset are FP16, the weights are  FP16, and the output is FP16. In the Interpreter kernels, each value from the embedding is dequantized to FP16, accumulation occurs in FP32, and then is casted back down to FP16 for the result.

For the Fused case, I added in a new `ElemKind::UInt8FusedFP16QTy`. This type has 4 extra bytes per row instead of 8, to save FP16 scale/offset.

Documentation: Updated.
Pull Request resolved: https://github.com/pytorch/glow/pull/3265

Test Plan: Added FP16 versions of all OperatorTests for RWQ-SLWS/SLS, fused and unfused.

Differential Revision: D16478496

Pulled By: jfix71

fbshipit-source-id: 9c46e4fea92ff8491093e9f341bd8b897b88164e