Add support for sqrt operator (#3306)
Summary:
**Summary**
This commit adds support for the elementwise sqrt operator to the
`PyTorchModelLoader` so that `torch.sqrt(a)` can be executed on Glow.

**Test Plan**
This commit adds a unit test that runs a graph with two sqrt nodes on Glow and
the PyTorch CPU backend and checks that the results are identical.
Pull Request resolved: https://github.com/pytorch/glow/pull/3306

Differential Revision: D16512707

Pulled By: SplitInfinity

fbshipit-source-id: 240d38f1706652a068ba95162838abb4305285e2