[PyTorch] Unbreak torch_glow
Summary:
torch_glow was inadvertently broken by #3308 which removed
`PyTorchModelLoader::populateNodeLoaderMapping` and created a static map
in its place. Support for exp, sqrt and max operators was added and
merged while PR #3308 was still a work in progress and it must have
gotten lost in the process of resolving merge conflicts.

Test Plan:
The unit tests for exp, sqrt and max pass again.