Fix memory usage when reusing allocator in compile (#3321)
Summary:
In the OpenCL backend we allocate all device buffers contiguously by reusing the same memory allocator for Constants, Placeholders and Activations. This diff fixes a bug in that logic which allocated the sum of the current stage and previous stages rather than just the memory needed for the current stage (e.g. we'd set sizes to be Constants, Placeholders + Constants, Activations + Placeholders + Constants).

This means we were allocating about 2-3x as much memory as we needed on the GPU. E.g. in OCLTest::ConvOps:

before:
```
  Constants: 2496
  Placeholders: 9216
  Activations: 12928
  Total alloced: 24640
```

after:
```
  Constants: 2496
  Placeholders: 6720
  Activations: 3712
  Total alloced: 12928
```

Gets better the bigger the Constants block is, obviously.

Documentation: n/a
Pull Request resolved: https://github.com/pytorch/glow/pull/3321

Test Plan: unit tests in ASAN and Release (OCLTest, BackendCorrectnessTest & OperatorTest being the important ones).

Differential Revision: D16565980

Pulled By: nickgg

fbshipit-source-id: 13130059fed41e308518c7ce69374039b245a82f