[PyTorch] Add support for avg_pool2d operator
Summary:
This commit adds support for the 2D average pooling operator to the
PyTorchModelLoader so that torch.nn.AvgPool2d can be executed on Glow.
This operator is required to support
torch.nn.LocalResponseNormalization and is also in resnet50.

Test Plan:
This commit adds a unit test that runs a graph with an 2D average pooling
node on Glow and the PyTorch CPU backend and checks that the results are
identical.