Fix OperatorTests::insertTensorTest (#3337)
Summary:
In OperatorTests::insertTensorTests the insertTensor is getting optimized out because all inputs are constant splats. Changing them to Placeholders not known at compile time fixes this.

Documentation: n/a
Pull Request resolved: https://github.com/pytorch/glow/pull/3337

Test Plan: unit tests, added printfs to libjit_insert_tensor to verify that the InsertTensorInst is in the final product.

Differential Revision: D16586565

Pulled By: nickgg

fbshipit-source-id: d7dc82ff6eaaf64d4382f66326b626a892596a56