[PyTorch] Allow caller of jitVsGlow to specify expected fused ops
Description:
This commit adds a parameter to jitVsGlow that the caller must use to
specify exactly which ops are expected to be placed into a Glow
compilation group. jitVsGlow checks that all ops outside of Glow
compilation groups are not in the list.

Testing:
All tests have been modified to use this new argument and they still
pass. After deleting a mapping in symbolLoaderMapping, the test
corresponding to that mapping fails.

Note that the transpose tests had to be disabled because the transpose
nodes don't get fused into the Glow fusion group for some reason.