Convert Placeholders to Constants (#3302)
Summary:
Allows each input of PyTorch nodes to be marked as inputs that should be preserved and converts all other inputs to glow Constants. This solves the problem that things such as conv weights are loaded as Placeholders because in PyTorch they may vary from one run to another. Glow however needs these to be Constants so that they can be offloaded to devices and run efficiently.

This is somewhat of medium-term hack and should be replaced eventually. The main issues with this are how manual marking each input is and that it doesn't do the right thing if the weights do actually change from one run to another because they will not get updated in Glow. A follow up to this PR should be to add a flag to totally enable or disable this conversion step.

Documentation:
Doxygen
Pull Request resolved: https://github.com/pytorch/glow/pull/3302

Test Plan:
`python setup.py test`
run resnet example and see that weights are now Constants.
Before
![Screen Shot 2019-07-25 at 3 51 23 PM](https://user-images.githubusercontent.com/1740091/61916833-08db6400-aeff-11e9-8e38-370436ee368a.png)

After
![Screen Shot 2019-07-25 at 3 50 07 PM](https://user-images.githubusercontent.com/1740091/61916837-0c6eeb00-aeff-11e9-8a24-0c86edb79a6f.png)

Differential Revision: D16642873

Pulled By: jackm321

fbshipit-source-id: 0a77e13525f27a270d5de45a6b7e986a20012712