Add support for avg_pool2d operator (#3346)
Summary:
**Summary:**
This commit adds support for the 2D average pooling operator to the
`PyTorchModelLoader` so that `torch.nn.AvgPool2d` can be executed on Glow.
This operator is required to support
`torch.nn.LocalResponseNormalization` and is also in resnet50.

**Test Plan:**
This commit adds a unit test that runs a graph with an 2D average pooling
node on Glow and the PyTorch CPU backend and checks that the results are
identical.
Pull Request resolved: https://github.com/pytorch/glow/pull/3346

Differential Revision: D16626569

Pulled By: SplitInfinity

fbshipit-source-id: 99a634da106128aab7378b8e07aca477ba72281b