Allow caller of jitVsGlow to specify expected fused ops (#3369)
Summary:
**Description:**
This commit adds a parameter to `jitVsGlow` that the caller must use to
specify exactly which ops are expected to be placed into a Glow
compilation group. `jitVsGlow` checks that all ops outside of Glow
compilation groups are not in the list.

**Test Plan:**
All tests have been modified to use this new argument and they still
pass. After deleting a mapping in `symbolLoaderMapping`, the test
corresponding to that mapping fails.

Note that the transpose tests had to be disabled because the transpose
nodes don't get fused into the Glow fusion group for some reason.
Pull Request resolved: https://github.com/pytorch/glow/pull/3369

Differential Revision: D16656778

Pulled By: SplitInfinity

fbshipit-source-id: f061b3dd2dce9859c548c04fbf4e4aec2acfdd43