DAG validation (#3404)
Summary:
This PR checks if a partition is a DAG.   This validation is mainly used in user-defined partition so far. But it can be added to anyplace where a partition DAG is already created.

Documentation:

[Optional Fixes #issue]
Pull Request resolved: https://github.com/pytorch/glow/pull/3404

Test Plan:
unit test added.

Please see a detailed explanation of how to fill out the fields in the relevant sections in PULL_REQUEST.md.

Reviewed By: gcatron

Differential Revision: D16745411

Pulled By: beicy

fbshipit-source-id: 5e29ae06bd63021eebc9c4fb65d5d55687d48909