Introduce Tensor alignments. (#3368)
Summary:
**Summary:**
In many architectures, data array is required (or recommended) to be aligned in memory. Meaning that total size of the array is multiple of X, and its starting memory address is a multiple of X. In case of multi-dimensional array (Tensor), each dimension should also be a multiple of X. It means that Tensor will have "holes" in it, i.e. useful data is padded by "garbage" values.

This PR introduces ability to create arbitrary-aligned Tensors, fixes various operations to work correctly.

**Documentation:**
- Class Type now has `strides_` member, which contains stride for each dimension.
- `Tensor::size()` still returns number of useful elements.
- `Tensor::actualSize()` is new method, returns number of allocated elements, including "holes" used for padding.
- `Tensor::size() <= Tensor::actualSize()` for every Tensor.
- We will not make any guarantees regarding actual value of pad-elements. They could be arbitrary numbers, all code should work correctly with any pads and not depend on their values.

**Test Plan:**
`ninja test`
Pull Request resolved: https://github.com/pytorch/glow/pull/3368

Differential Revision: D16617267

Pulled By: artemrakhov

fbshipit-source-id: b0c73c8e10da62780e4c1c70532ee1747032324b